refactor(tests): reorganize test modules and remove unused files
