+++ .github/workflows/continuous-integration.yml
@@ -71,6 +71,6 @@ jobs:
       - name: Format Check
         run: cargo fmt -- --check
       - name: Clippy Check
-        run: cargo clippy --release -- -D warnings
+        run: cargo clippy --release --workspace -- -D warnings
       - name: Test
-        run: cargo test --release
+        run: cargo test --release --workspace
