+++ crates/server/src/tests/middleware/check_auth.rs
@@ -1,7 +1,8 @@
 #[cfg(test)]
 mod tests {
-    use crate::{settings::Settings, utils::check_auth::check_auth};
+    use crate::middleware::auth::check_auth;
     use actix_web_httpauth::{extractors::basic::BasicAuth, headers::authorization::Basic};
+    use autopulse_service::settings::Settings;
 
     #[test]
     fn test_check_default_auth() -> anyhow::Result<()> {
@@ -11,7 +12,12 @@ mod tests {
         ));
         let settings: Settings = serde_json::from_str("{}")?;
 
-        assert!(check_auth(&Some(auth), &settings));
+        assert!(check_auth(
+            &Some(auth),
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
@@ -21,7 +27,12 @@ mod tests {
         let auth = BasicAuth::from(Basic::new("username".to_string(), Some(String::new())));
         let settings: Settings = serde_json::from_str("{}")?;
 
-        assert!(!check_auth(&Some(auth), &settings));
+        assert!(!check_auth(
+            &Some(auth),
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
@@ -34,7 +45,12 @@ mod tests {
         ));
         let settings: Settings = serde_json::from_str("{\"auth\":{\"username\":\"username\"}}")?;
 
-        assert!(check_auth(&Some(auth), &settings));
+        assert!(check_auth(
+            &Some(auth),
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
@@ -44,7 +60,12 @@ mod tests {
         let auth = BasicAuth::from(Basic::new("admin".to_string(), Some("pass".to_string())));
         let settings: Settings = serde_json::from_str("{\"auth\":{\"password\":\"pass\"}}")?;
 
-        assert!(check_auth(&Some(auth), &settings));
+        assert!(check_auth(
+            &Some(auth),
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
@@ -54,7 +75,12 @@ mod tests {
         let auth = BasicAuth::from(Basic::new("admin".to_string(), Some("pass".to_string())));
         let settings: Settings = serde_json::from_str("{\"auth\":{\"enabled\": false}}")?;
 
-        assert!(check_auth(&Some(auth), &settings));
+        assert!(check_auth(
+            &Some(auth),
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
@@ -63,7 +89,12 @@ mod tests {
     fn test_check_disabled_auth_none() -> anyhow::Result<()> {
         let settings: Settings = serde_json::from_str("{\"auth\":{\"enabled\": false}}")?;
 
-        assert!(check_auth(&None, &settings));
+        assert!(check_auth(
+            &None,
+            &settings.auth.enabled,
+            &settings.auth.username,
+            &settings.auth.password
+        ));
 
         Ok(())
     }
