+++ crates/service/src/lib.rs
@@ -5,3 +5,8 @@ pub mod runner;
 
 /// Settings for the service
 pub mod settings;
+
+#[cfg(test)]
+mod tests {
+    mod triggers;
+}
