+++ crates/service/src/settings/targets/autopulse.rs
@@ -1,7 +1,7 @@
 use crate::settings::rewrite::Rewrite;
 use crate::settings::{auth::Auth, targets::TargetProcess};
 use autopulse_database::models::ScanEvent;
-use autopulse_utils::get_url::get_url;
+use autopulse_utils::get_url;
 use reqwest::header;
 use serde::Deserialize;
 use tracing::{debug, error};
