+++ crates/service/src/settings/triggers/radarr.rs
@@ -1,7 +1,7 @@
 use crate::settings::rewrite::Rewrite;
 use crate::settings::timer::Timer;
 use crate::settings::triggers::TriggerRequest;
-use autopulse_utils::join_path::join_path;
+use autopulse_utils::join_path;
 use serde::Deserialize;
 
 #[derive(Deserialize, Clone)]
