+++ crates/utils/src/tests/mod.rs
@@ -1,4 +1,3 @@
-pub mod check_auth;
 pub mod checksum;
 pub mod generate_uuid;
 pub mod get_url;
